move non-sampling files from sampling directory to a new directory
