+++ src/utilities/output_quantities/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(${kynema_sgf_lib_name}
+  PRIVATE
+
+  Enstrophy.cpp
+  FieldNorms.cpp
+  KineticEnergy.cpp
+  WaveEnergy.cpp
+  )
