+++ src/utilities/output_quantities/WaveEnergy.H
@@ -10,7 +10,7 @@ using namespace amrex::literals;
 namespace kynema_sgf::wave_energy {
 
 /** wave energy object
- *  \ingroup sampling
+ *  \ingroup output_quantities
  *
  *  A concrete implementation of the post-processing interface that deals with
  *  integrating total wave energy. This routine calculates the kinetic energy
