+++ unit_tests/utilities/test_wave_energy.cpp
@@ -1,6 +1,6 @@
 #include <utility>
 #include "ks_test_utils/MeshTest.H"
-#include "src/utilities/sampling/WaveEnergy.H"
+#include "src/utilities/output_quantities/WaveEnergy.H"
 #include "src/utilities/math_ops.H"
 
 using namespace amrex::literals;
